perf(views): [large-screen/wisdom-tourism] 优化大屏看板/智慧旅游模块 - 增加创建缩放比例的容器的高阶函数、优化按钮样式
