+++ RECORD.md
@@ -74,6 +74,7 @@ Refs: #123
 - fix(build): 修复 CI/CD 构建失败错误问题
 - chore(ci/cd): 更改ci/cd运行的分支
 - fix(eslint): 修复 ci/cd 构建时触发 eslint 校验错误问题
+- perf(views): [large-screen/wisdom-tourism] 优化大屏看板/智慧旅游模块 - 增加创建缩放比例的容器的高阶函数、优化按钮样式
 
 ## 2024年10月9日-2024年10月10日
