+++ src/index.scss
@@ -3,6 +3,8 @@
   color: rgb(255 255 255 / 87%);
   color-scheme: light dark;
   font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
+
+  // 抗锯齿技术，增加字体清晰度
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
   font-synthesis: none;
